fix[Op#40437]: Reinstate emoji reactions streams
Journalsa are updated on last updated_at; reactions cannot rely on this as they do not "touch"
the journal, so a full journal refresh would not be viable. We avoid "touching" the journal on emoji
reactions update to retain concurrent user actions. E.g. If a user has a journal in edit mode, we do not want to overrite
it on the next polling update- this is why it's safer to stream emoji reactions separately

This reverts commit fb419dd23be7aae2a4dc9e5e4fa6c08ccd23889e.