Add CAPYBARA_PUMA_THREADS to control puma threads count
When using `binding.irb` in backend code, we can have multiple REPL
spawning if multiple requests are made. This makes it hard to use IRB as
it will jump from one thread to another each time an expression is
evaluated. Barely usable.

This commit adds a new environment variable `CAPYBARA_PUMA_THREADS` to
control the number of threads Puma uses in tests. Set it to "1:1" to use
only one thread, which is useful when using IRB. Not setting it will
continue to use the default number of threads for Puma in test
environment.

See https://github.com/puma/puma/blob/master/lib/rack/handler/puma.rb#L7
for available options.